added ValueContainer and AttachableContainer extension interface
AttachableContainer is needed by ObjectGraph to be able to pull
dependencies from its supercontainer in a ContainerGraph.

i also have the option of making the supercontainer able to pull deps
from the subcontainer, not sure if i'm going with it. that is opposed to
the semantics of dagger so i don't know if there are problems with it.